Added new ramps
added a new ramp towards the new player base location and updated the GBI to show it.
also changed the terrain of one part of the hill to allow for a second ramp towards the
Gensher's base, so now there is a safe route between the two bases.

His waypoints still go through the death canyon though.