Mission plays pretty well and nice
the debug logs added to pinpoint the pathing issue worked, but didn't bring
anything new to the table since the pathing managed to find a way this time
and didn't get stuck.

in any case, put the debug log behind a 'logsenabled' variable.

The only gripe i still have is that Khattam becomes an arab rather than staying an american.